essi-oauth/requirements.txt
flask-lambda==0.0.4
werkzeug==2.0.3
requests==2.31.0
bcrypt==4.0.1
pyjwt==2.8.0